extract event details component
